PythonML/requirements.txt
# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Core ML Libraries
torch==2.1.2
transformers==4.36.0
accelerate==0.25.0
sentencepiece==0.1.99
numpy==1.24.3
Pillow==10.1.0

# Specialized Pipelines
sentence-transformers==2.2.2
soundfile==0.12.1
opencv-python==4.8.1.78
scipy==1.11.4
einops==0.7.0
xformers==0.0.23
bitsandbytes==0.41.3
optimum==1.16.1
diffusers==0.25.0
audiocraft==1.3.0
transformers_stream_generator==0.0.5

# MediaPipe (GenAI + Vision + Text + Audio)
mediapipe==0.10.9

# LiteRT (TensorFlow Lite)
tensorflow==2.15.0

# SpaCy (NER, optional)
spacy==3.7.2

# System utilities
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
